relax /system/bin directory permissions
In commit 31cdadb65e142522a0eda83e3d70abb315a9e584 [formerly f4fc922f0b863659ca8e97c1f5fa522fafc7deb6], we tightened the
permissions on various bin directories. Please see
https://android-review.googlesource.com/c/platform/system/core/+/822955
for details.

This change causes the Chase banking app to crash. This is because
the Chase app is using inotify_add_watch() on the /system/bin directory
and not checking the return value.

The Android Security model guarantees the immutability of files in
/system/bin, so the inotify watch is unnecessary.

Until the Chase app fixes their bug, we need to relax the permissions on
the /system/bin directory. Conceptually, this is a partial revert of
31cdadb65e142522a0eda83e3d70abb315a9e584 [formerly f4fc922f0b863659ca8e97c1f5fa522fafc7deb6].

Bug: 119605322
Test: compiles
Change-Id: Ic72dd24cb27cff677093963bdfd0ae09bf132e08

Former-commit-id: 53842f8a9044db84109ee84dbacef09491be15c2